refactor(cf): relocate cf-app-variables types with its signal config
