+++ src/frontend/packages/cloud-foundry/src/features/applications/application/application-tabs-base/tabs/variables-tab/variables-tab.component.ts
@@ -26,7 +26,7 @@ import {
 import { CodeBlockComponent } from '../../../../../../../../core/src/shared/components/code-block/code-block.component';
 import {
   ListAppEnvVar,
-} from '../../../../../../shared/components/list/list-types/app-variables/cf-app-variables.types';
+} from '../../../../../../shared/signal-list-configs/app-variables/cf-app-variables.types';
 import {
   CfAppVariablesSignalConfigService,
 } from '../../../../../../shared/signal-list-configs/app-variables/cf-app-variables-signal-config.service';
