+++ src/frontend/packages/cloud-foundry/src/shared/signal-list-configs/app-variables/cf-app-variables.types.ts
